support specifying prior in config
